pyproject.toml
[tool.poetry]
name = "clypper"
version = "0.0.1"
description = "Rapidly create supercuts from various video sources."
authors = ["kpj <kpjkpjkpjkpjkpjkpj@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.1.2"
sh = "^1.13.1"
pytube3 = "^9.6.4"
moviepy = "^1.0.3"
tqdm = "^4.46.0"
pytimeparse = "^1.1.8"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
tox = "^3.15.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
